refactor: update package references to use sharedinfra namespace
