+++ boot/src/main/java/pt/estga/boot/config/AuditingConfig.java
@@ -4,7 +4,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.domain.AuditorAware;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-import pt.estga.shared.models.AuditActor;
+import pt.estga.sharedinfra.models.AuditActor;
 import pt.estga.sharedcore.utils.SecurityUtils;
 
 @Configuration
