+++ file/src/main/java/pt/estga/file/entities/MediaFile.java
@@ -8,7 +8,7 @@ import lombok.*;
 import org.hibernate.annotations.UuidGenerator;
 import pt.estga.file.enums.MediaStatus;
 import pt.estga.file.enums.StorageProvider;
-import pt.estga.shared.entities.CreationAuditedEntity;
+import pt.estga.sharedinfra.entities.CreationAuditedEntity;
 
 import java.util.ArrayList;
 import java.util.List;
