+++ intake/src/main/java/pt/estga/intake/services/SubmissionQueryService.java
@@ -11,7 +11,7 @@ import pt.estga.intake.entities.MarkEvidenceSubmission;
 import pt.estga.intake.enums.SubmissionStatus;
 import pt.estga.intake.mappers.SubmissionMapper;
 import pt.estga.intake.repositories.MarkEvidenceSubmissionRepository;
-import pt.estga.shared.jpa.SpecBuilder;
+import pt.estga.sharedinfra.jpa.SpecBuilder;
 
 import java.util.Optional;
