+++ mark/src/main/java/pt/estga/mark/entities/Mark.java
@@ -2,8 +2,8 @@ package pt.estga.mark.entities;
 
 import jakarta.persistence.*;
 import lombok.*;
-import pt.estga.shared.converters.ValidationStateConverter;
-import pt.estga.shared.entities.BaseEntity;
+import pt.estga.sharedinfra.converters.ValidationStateConverter;
+import pt.estga.sharedinfra.entities.BaseEntity;
 import pt.estga.sharedcore.enums.ValidationState;
 
 import java.util.List;
