+++ mark/src/main/java/pt/estga/mark/repositories/MarkRepository.java
@@ -3,7 +3,7 @@ package pt.estga.mark.repositories;
 import org.springframework.stereotype.Repository;
 import pt.estga.mark.entities.Mark;
 import pt.estga.sharedcore.enums.ValidationState;
-import pt.estga.shared.repositories.BaseRepository;
+import pt.estga.sharedinfra.repositories.BaseRepository;
 
 import java.util.List;
