+++ shared-infra/src/main/java/pt/estga/sharedinfra/entities/CreationAuditedEntity.java
@@ -1,4 +1,4 @@
-package pt.estga.shared.entities;
+package pt.estga.sharedinfra.entities;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
@@ -9,7 +9,7 @@ import lombok.experimental.SuperBuilder;
 import org.springframework.data.annotation.CreatedBy;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-import pt.estga.shared.models.AuditActor;
+import pt.estga.sharedinfra.models.AuditActor;
 
 import java.time.Instant;
