+++ shared-infra/src/main/java/pt/estga/sharedinfra/repositories/BaseRepository.java
@@ -1,8 +1,8 @@
-package pt.estga.shared.repositories;
+package pt.estga.sharedinfra.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.NoRepositoryBean;
-import pt.estga.shared.entities.BaseEntity;
+import pt.estga.sharedinfra.entities.BaseEntity;
 
 @NoRepositoryBean
 public interface BaseRepository<T, ID> extends JpaRepository<T, ID> {
