+++ support/src/main/java/pt/estga/support/services/ContactRequestService.java
@@ -5,7 +5,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import pt.estga.shared.jpa.SpecBuilder;
+import pt.estga.sharedinfra.jpa.SpecBuilder;
 import pt.estga.sharedweb.exceptions.ResourceNotFoundException;
 import pt.estga.support.dtos.ContactRequestDto;
 import pt.estga.support.dtos.ContactRequestFilter;
