+++ territory/src/main/java/pt/estga/territory/repositories/AdministrativeDivisionRepository.java
@@ -1,11 +1,10 @@
 package pt.estga.territory.repositories;
 
-import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
-import pt.estga.shared.repositories.BaseRepository;
+import pt.estga.sharedinfra.repositories.BaseRepository;
 import pt.estga.territory.entities.AdministrativeDivision;
 
 import java.util.List;
