+++ territory/src/main/java/pt/estga/territory/repositories/CountryRepository.java
@@ -1,6 +1,6 @@
 package pt.estga.territory.repositories;
 
-import pt.estga.shared.repositories.BaseRepository;
+import pt.estga.sharedinfra.repositories.BaseRepository;
 import pt.estga.territory.entities.Country;
 import java.util.Optional;
